Add support for loading linux-related images
This adds support for loading linux images using the [load]
attribute. The [offset] attribute is not yet supported.

The program will try to recognize to file format, but if the format
is not recognized, it will be put into the image as a whole without
setting any additional attribute bits in the partition header.